# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Auth / UserManager",
    "version": "1.0.0",
    "description": "Documentation de l'API Auth / UserManager (SIWE, OTP, JWT, KYC)"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Serveur local"
    }
  ],
  "paths": {
    "/api/treasury/balance": {
      "get": {
        "summary": "Consulter le solde de la trésorerie",
        "tags": [
          "Treasury"
        ],
        "responses": {
          "200": {
            "description": "Solde de la trésorerie récupéré avec succès"
          },
          "400": {
            "description": "Erreur lors de la récupération du solde"
          }
        }
      }
    },
    "/api/treasury/withdraw": {
      "post": {
        "summary": "Retirer des fonds depuis la trésorerie (admin seulement)",
        "tags": [
          "Treasury"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to",
                  "amount"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "example": "0xAbC123456789..."
                  },
                  "amount": {
                    "type": "string",
                    "example": "1000000000000000000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retrait effectué avec succès"
          },
          "400": {
            "description": "Erreur ou autorisation refusée"
          }
        }
      }
    },
...
    "/api/wallets/rate/{currency}": {
      "get": {
        "tags": [
          "Admin Wallets"
        ],
        "summary": "Récupérer le taux de change d'une devise",
        "description": "Retourne le taux de change configuré sur le contrat `MultiWalletExchange`.\n",
        "operationId": "getExchangeRate",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "currency",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EUR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Taux récupéré"
          },
          "404": {
            "description": "Devise inconnue"
          },
          "500": {
            "description": "Erreur blockchain"
          }
        }
      }
    },
    "/api/wallets/rate": {
      "post": {
        "tags": [
          "Admin Wallets"
        ],
        "summary": "Définir un taux de change -> C'EST OK",
        "description": "Configure le taux de change d'une devise dans le contrat `MultiWalletExchange`.\n**⚠️ Réservé aux administrateurs (owner du contrat)**.\n",
        "operationId": "setExchangeRate",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currency",
                  "rate"
                ],
                "properties": {
                  "currency": {
                    "type": "string",
                    "example": "EUR"
                  },
                  "rate": {
                    "type": "string",
                    "example": "1200"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Taux mis à jour"
          },
          "401": {
            "description": "Non autorisé"
          },
          "500": {
            "description": "Erreur blockchain"
          }
        }
      }
    },
    "/api/wallets/spread/{currency}": {
      "get": {
        "tags": [
          "Admin Wallets"
        ],
        "summary": "Récupérer le spread appliqué à une devise",
        "description": "Retourne le spread (marge) appliqué pour les conversions d'une devise donnée.\n",
        "operationId": "getSpread",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "currency",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EUR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spread récupéré"
          },
          "404": {
            "description": "Devise inconnue"
          },
          "500": {
            "description": "Erreur blockchain"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtenu lors du login"
      }
    }
  },
...
}